backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
psycopg2-binary>=2.9.9
httpx>=0.24.0
razorpay>=1.4.1
apscheduler>=3.10.4
python-multipart>=0.0.9
